signals & heredoc is fixed

+++ libft/libft.h
@@ -6,13 +6,14 @@
 /*   By: momeaizi <momeaizi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/01 11:16:59 by momeaizi          #+#    #+#             */
-/*   Updated: 2022/06/29 11:56:10 by momeaizi         ###   ########.fr       */
+/*   Updated: 2022/07/02 12:06:06 by momeaizi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFT_H
 # define LIBFT_H
 
+# include <limits.h>
 # include <errno.h>
 # include <fcntl.h>
 # include <signal.h>
@@ -27,43 +28,43 @@ typedef struct s_list
 	struct s_list	*next;
 }	t_list;
 
-char	*ft_strnstr(const char *haystack, const char *needle, size_t len);
-char	*ft_substr(char const *s, unsigned int start, size_t len);
-size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize);
-size_t	ft_strlcat(char *dst, const char *src, size_t dstsize);
-int		ft_strncmp(const char *s1, const char *s2, size_t n);
-void	*ft_memmove(void *dst, const void *src, size_t len);
-int		ft_memcmp(const void *s1, const void *s2, size_t n);
-void	*ft_memcpy(void *dst, const void *src, size_t n);
-char	*ft_strtrim(char const *s1, char const *set);
-char	*ft_strjoin(char const *s1, char const *s2);
-void	*ft_memchr(const void *s, int c, size_t n);
-int		skip_quotes(char *line, int i, char quote);
-void	*ft_memset(void *b, int c, size_t len);
-void	*ft_calloc(size_t count, size_t size);
-char	**ft_realloc(char **strs, char *str);
-char	*ft_strrchr(const char *s, int c);
-char	*ft_strchr(const char *s, int c);
-void	ft_putendl_fd(char *s, int fd);
-int		skip_space(char *line, int i);
-int		ft_strcmp(char *s1, char *s2);
-void	ft_putchar_fd(char c, int fd);
-void	ft_putstr_fd(char *s, int fd);
-void	ft_bzero(void *s, size_t n);
-void	ft_putnbr_fd(int n, int fd);
-char	*ft_strdup(const char *s1);
-int		ft_atoi(const char *str);
-int		size_double(char **cmds);
-size_t	ft_strlen(const char *s);
-size_t	ft_strlen(const char *s);
-char	*add_spaces(char *line);
-int		ft_isalpha(int c);
-int		ft_isdigit(int c);
-int		ft_isalnum(int c);
-int		ft_isascii(int c);
-int		ft_isprint(int c);
-int		ft_toupper(int c);
-int		ft_tolower(int c);
-char	*ft_itoa(int n);
+char		*ft_strnstr(const char *haystack, const char *needle, size_t len);
+char		*ft_substr(char const *s, unsigned int start, size_t len);
+size_t		ft_strlcpy(char *dst, const char *src, size_t dstsize);
+size_t		ft_strlcat(char *dst, const char *src, size_t dstsize);
+int			ft_strncmp(const char *s1, const char *s2, size_t n);
+void		*ft_memmove(void *dst, const void *src, size_t len);
+int			ft_memcmp(const void *s1, const void *s2, size_t n);
+void		*ft_memcpy(void *dst, const void *src, size_t n);
+char		*ft_strtrim(char const *s1, char const *set);
+char		*ft_strjoin(char const *s1, char const *s2);
+void		*ft_memchr(const void *s, int c, size_t n);
+int			skip_quotes(char *line, int i, char quote);
+void		*ft_memset(void *b, int c, size_t len);
+void		*ft_calloc(size_t count, size_t size);
+char		**ft_realloc(char **strs, char *str);
+char		*ft_strrchr(const char *s, int c);
+char		*ft_strchr(const char *s, int c);
+void		ft_putendl_fd(char *s, int fd);
+int			skip_space(char *line, int i);
+int			ft_strcmp(char *s1, char *s2);
+void		ft_putchar_fd(char c, int fd);
+void		ft_putstr_fd(char *s, int fd);
+void		ft_bzero(void *s, size_t n);
+void		ft_putnbr_fd(int n, int fd);
+char		*ft_strdup(const char *s1);
+long long	ft_atoi(const char *str);
+int			size_double(char **cmds);
+size_t		ft_strlen(const char *s);
+size_t		ft_strlen(const char *s);
+char		*add_spaces(char *line);
+int			ft_isalpha(int c);
+int			ft_isdigit(int c);
+int			ft_isalnum(int c);
+int			ft_isascii(int c);
+int			ft_isprint(int c);
+int			ft_toupper(int c);
+int			ft_tolower(int c);
+char		*ft_itoa(int n);
 
 #endif
